+++ CHANGES.md
@@ -1,5 +1,13 @@
 # Change Log
 
+# 0.3.0
+
+- Added full complement of 0.3.0 features
+  - vRouter support
+  - Base64 enable and passwords
+  - Configuration groups (i.e. one for core/edge/default)
+  - Some other minor bug fixes
+
 # 0.2.0
 
 - Update `config.schema.yaml` to treat enable password as secret, and connection method as enum